Clear up some TODOs
