+++ code/drasil-code/Language/Drasil/Chunk/CodeDefinition.hs
@@ -23,35 +23,33 @@ data CodeDefinition = CD { _cchunk   :: CodeChunk
 makeLenses ''CodeDefinition
 
 -- | Finds the 'UID' of the 'CodeChunk' used to make the 'CodeDefinition'.
-instance HasUID       CodeDefinition where uid = cchunk . uid
+instance HasUID           CodeDefinition where uid = cchunk . uid
 -- | Finds the term ('NP') of the 'CodeChunk' used to make the 'CodeDefinition'.
-instance NamedIdea    CodeDefinition where term = cchunk . term
+instance NamedIdea        CodeDefinition where term = cchunk . term
 -- | Finds the idea contained in the 'CodeChunk' used to make the 'CodeDefinition'.
-instance Idea         CodeDefinition where getA = getA . view cchunk
+instance Idea             CodeDefinition where getA = getA . view cchunk
 -- | Finds the 'Space' of the 'CodeChunk' used to make the 'CodeDefinition'.
-instance HasSpace     CodeDefinition where typ = cchunk . typ
+instance HasSpace         CodeDefinition where typ = cchunk . typ
 -- | Finds the 'Stage' dependent 'Symbol' of the 'CodeChunk' used to make the 'CodeDefinition'.
-instance HasSymbol    CodeDefinition where symbol c = symbol (c ^. cchunk)
+instance HasSymbol        CodeDefinition where symbol c = symbol (c ^. cchunk)
 -- | 'CodeDefinition's have a 'Quantity'.
-instance Quantity     CodeDefinition
+instance Quantity         CodeDefinition
 -- | Finds the code name of a 'CodeDefinition'.
 -- 'Function' 'CodeDefinition's are named with the function prefix to distinguish 
 -- them from the corresponding variable version.
-instance CodeIdea     CodeDefinition where 
+instance CodeIdea         CodeDefinition where 
   codeName (CD c@(CodeC _ Var) _ _ _) = codeName c
   codeName (CD c@(CodeC _ Func) _ _ _) = funcPrefix ++ codeName c
   codeChunk = view cchunk
 -- | Equal if 'UID's are equal.
-instance Eq           CodeDefinition where c1 == c2 = (c1 ^. uid) == (c2 ^. uid)
+instance Eq               CodeDefinition where c1 == c2 = (c1 ^. uid) == (c2 ^. uid)
 -- | Finds the units of the 'CodeChunk' used to make the 'CodeDefinition'.
-instance MayHaveUnit  CodeDefinition where getUnit = getUnit . view cchunk
+instance MayHaveUnit      CodeDefinition where getUnit = getUnit . view cchunk
 -- | Finds the defining expression of a CodeDefinition.
 instance DefiningCodeExpr CodeDefinition where codeExpr = def
--- instance DefiningExpr CodeDefinition where defnExpr = def
 
-
--- TODO: These below smart constructors will surely need to be able to use custom
---       code expressions, since they currently only allow deriving via normal Expr.
+-- NOTE: We'll eventually want extra smart constructors that allow for custom
+--       CodeExprs inputs.
 
 -- | Constructs a 'CodeDefinition' where the underlying 'CodeChunk' is for a function.
 qtoc :: (Quantity q, DefiningExpr q, MayHaveUnit q) => q -> CodeDefinition
@@ -61,7 +59,6 @@ qtoc q = CD (codeChunk $ quantfunc q) (renderExpr $ q ^. defnExpr) [] Definition
 qtov :: QDefinition -> CodeDefinition
 qtov q = CD (codeChunk $ quantvar q) (renderExpr $ q ^. defnExpr) [] Definition
 
--- TODO: ODEInfos should probably be holding CodeExprs on their own, without us rendering it here
 -- | Constructs a 'CodeDefinition' for an ODE.
 odeDef :: ODEInfo -> CodeDefinition
 odeDef info = CD (codeChunk $ quantfunc $ depVar info) (matrix [map renderExpr $ odeSyst info])
