+++ code/drasil-printers/Language/Drasil/Printing/Import/Symbol.hs
@@ -1,10 +1,13 @@
-module Language.Drasil.Printing.Import.Symbol where
+module Language.Drasil.Printing.Import.Symbol (symbol, pUnit) where
 
+import Language.Drasil (USymb(..))
 import Language.Drasil.Display (Decoration(..), Symbol(..))
 import qualified Language.Drasil.Printing.AST as P
 
+import Data.List (partition)
+import Data.Bifunctor (second)
 
--- | Helper tha converts a symbol into an expression.
+-- | Converts a symbol into an expression.
 symbol :: Symbol -> P.Expr
 symbol (Variable s) = P.Ident s
 symbol (Label    s) = P.Label s
@@ -13,7 +16,7 @@ symbol (Special  s) = P.Spec s
 --symbol (Greek g)    = P.Gr g
 symbol (Concat  sl) = P.Row $ map symbol sl
 --
--- handle the special cases first, then general case
+-- Handle the special cases first, then general case
 symbol (Corners [] [] [x] [] s) = P.Row [P.Row [symbol s, P.Sup $ symbol x]]
 symbol (Corners [] [] [] [x] s) = P.Row [P.Row [symbol s, P.Sub $ symbol x]]
 symbol (Corners [_] [] [] [] _) = error "rendering of ul prescript"
@@ -27,3 +30,20 @@ sFormat :: Decoration -> Symbol -> P.Expr
 sFormat Hat    s = P.Over P.Hat $ symbol s
 sFormat Vector s = P.Font P.Bold $ symbol s
 sFormat Prime  s = P.Row [symbol s, P.MO P.Prime]
+
+-- | Renders a unit symbol as a printable expression.
+pUnit :: USymb -> P.Expr
+pUnit (US ls) = formatu t b
+  where
+    (t, b) = partition ((> 0) . snd) ls
+    formatu :: [(Symbol, Integer)] -> [(Symbol, Integer)] -> P.Expr
+    formatu [] l = line l
+    formatu l [] = P.Row $ map powu l
+    formatu nu de = P.Div (line nu) $ line $ map (second negate) de
+    line :: [(Symbol, Integer)] -> P.Expr
+    line []  = P.Row [] -- should not happen ?
+    line [n] = powu n
+    line l   = P.Row $ map powu l
+    powu :: (Symbol, Integer) -> P.Expr
+    powu (n, 1) = symbol n
+    powu (n, p) = P.Row [symbol n, P.Sup $ P.Int p]
