+++ code/drasil-printers/drasil-printers.cabal
@@ -4,7 +4,7 @@ cabal-version: 1.12
 --
 -- see: https://github.com/sol/hpack
 --
--- hash: 7e8fb691167d024073cb90d05bfc003b88172f47131ce85cee2e6b0cb5e67832
+-- hash: 8910014a28a0a4b118d7c4fdb876503418ed068f6a695793b86b282218274bc2
 
 name:           drasil-printers
 version:        0.1.10.0
@@ -38,8 +38,10 @@ library
       Language.Drasil.Printing.Import
       Language.Drasil.Printing.Import.CodeExpr
       Language.Drasil.Printing.Import.DisplayExpr
+      Language.Drasil.Printing.Import.Document
       Language.Drasil.Printing.Import.Expr
       Language.Drasil.Printing.Import.Helpers
+      Language.Drasil.Printing.Import.Sentence
       Language.Drasil.Printing.Import.Space
       Language.Drasil.Printing.Import.Symbol
       Language.Drasil.Printing.LayoutObj
